+++ quantforge/__init__.py
@@ -534,6 +534,7 @@ from .dual_calculus import dual_gradient, dual_newton
 from .hyperdual import HyperDual, derivatives as hyperdual_derivatives, second_derivative
 from .logsumexp import logsumexp, softmax, log_softmax
 from .kde import kde, silverman_bandwidth, scott_bandwidth, kde_function
+from .regression_tree import fit_regression_tree, predict_regression_tree
 from .levenberg import levenberg_marquardt
 from .ewma_cov import ewma_covariance_matrix, ewma_correlation_matrix
 from .vol_convert import black_to_normal_vol, normal_to_black_vol
@@ -924,7 +925,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "1.961.0"
+__version__ = "1.962.0"
 
 __all__ = [
     "OptionType",
@@ -1719,6 +1720,8 @@ __all__ = [
     "silverman_bandwidth",
     "scott_bandwidth",
     "kde_function",
+    "fit_regression_tree",
+    "predict_regression_tree",
     "levenberg_marquardt",
     "ewma_covariance_matrix",
     "ewma_correlation_matrix",